- updated table struct - updated README - comment some features

+++ models/messages.js
@@ -1,12 +1,12 @@
 'use strict';
 module.exports = (sequelize, DataTypes) => {
     var Messages = sequelize.define('Messages', {
-        message_id: DataTypes.INTEGER,
-        user_id: DataTypes.INTEGER,
+        message_id: DataTypes.BIGINT,
+        user_id: DataTypes.BIGINT,
         first_name: DataTypes.STRING,
         last_name: DataTypes.STRING,
         username: DataTypes.STRING,
-        chat_id: DataTypes.INTEGER,
+        chat_id: DataTypes.BIGINT,
         chat_title: DataTypes.STRING,
         chat_type: DataTypes.STRING,
         message: DataTypes.TEXT,
