+++ models/reminder.js
@@ -1,7 +1,7 @@
 'use strict';
 module.exports = (sequelize, DataTypes) => {
   var Reminder = sequelize.define('Reminder', {
-    userId: DataTypes.INTEGER,
+    userId: DataTypes.BIGINT,
     reminder: DataTypes.STRING,
     end: DataTypes.DATE,
     status: DataTypes.INTEGER
